Added functionality for Drift Correction Changes. The testing points were converted to drift correction points. Now when you add a drift correction point, all subsequent frames are automatically corrected for the drift in the glasses.
The primary and secondary calibrations are not adjusted for drift when plotted, but the output in the .txt file is corrected for drift.

This will mean that users should no longer use the linearcalibration, but instead always use the primary and use drift correction to fix problems that happen throughout the experiment.